pyproject.toml
[project]
name = "docker-volume-backaper"
version = "0.1.0"
description = "Backuper for remote docker volumes"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "paramiko>=3.5.1",
    "pyyaml>=6.0.2",
    "scp>=0.15.0",
    "tqdm>=4.67.1",
    "types-pyyaml>=6.0.12.20250402",
    "tzlocal>=5.3.1",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.9",
]

[tool.ruff]
line-length = 120
lint.select = ["ALL"]
lint.ignore = ["D100", "D101", "D102", "D103", "D104", "D107", "D203", "D211", "D213", "COM812", "S603"]
target-version = "py313"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
